Modify setting files
