--- configs/training/default.yaml
@@ -1,23 +1,0 @@
-PREPROCESS:
-  WIDTH: 512
-  HEIGHT: 512
-  OVERLAP: 0.2
-
-DATASET:
-  TRAIN:
-    IMAGE: "/home/ubuntu/workspace/ship_detection/dataset/JPEGImages_sub"
-    ANNOTATION: "/home/ubuntu/workspace/ship_detection/dataset/train.json"
-  VAL:
-    IMAGE: "/home/ubuntu/workspace/ship_detection/dataset/JPEGImages_sub"
-    ANNOTATION: "/home/ubuntu/workspace/ship_detection/dataset/val.json"
-  NUM_CLASS: 2
-
-TRAIN:
-  EPOCH: 20
-  BATCH_SIZE: 4
-  OPTIMIZER:
-    LEARNING_RATE: 0.0001
-    MOMENTUM: 0.9
-    WEIGHT_DECAY: 0.0005
-    LR_STEP_SIZE: 3
-    LR_GAMMA: 0.1
