+++ configs/yolof/yolof_r50_c5_8x8_1x_coco.py
